Refactor proto endpoint to use IncomingConnection
This commit refactors the logic for a quinn_proto::Endpoint accepting
an incoming connection so that it constructs an explicit
IncomingConnection struct containing all the necessary state to accept/
reject/retry the connection-creating packet. However, the external API
stays the same.

The bulk of this code change is just moving around existing code.

Also adds some gitignore lines I was using for coverage testing.