Allow accept/reject/retry before handshake begins
This commit removes use_retry from the server config and provides a
public API for the user to manually accept/reject/retry incoming
connections before a handshake begins, and inspect properties such as
an incoming connection's remote address and whether that address is
validated when doing so.

In quinn-proto, IncomingConnection is made public, as well as
Endpoint's accept/reject/retry methods which operate on it. The
DatagramEvent::NewConnection event is modified to return an incoming
but not yet accepted connection.

In quinn, Connecting is given two new internal states: Incoming,
representing an incoming but not yet accepted connection, and
AcceptError, which is necessary due to the fact that it is now possible
for errors to occur in the processing of a connection-creating packet
after it was already presented to the application as a Connecting.

Incoming Connecting futures are now yielded to the application to
represent potential incoming connections for which the server has not
yet sent a reply, allocated state, or otherwise committed to accepting
it. Existing functionality can mostly be used as normal, and operations
such as polling will automatically trigger the accepting of the
connection. However, methods is_validated, may_retry, reject, and retry
are added also, and existing method remote_address is made to work
without/before accepting.

One notable case of backwards-incompatibility is that it is now
possible for into_0rtt to error for incoming connections, which was not
previously the case. The error type of into_0rtt is changed into a new
error type which intentionally avoids implementing Debug so as to cause
user code which was blindly unwrapping such errors in servers to fail
to compile.

The test accept_after_close was removed because the functionality it
was testing for no longer exists.