Add TokenMemoryCache
TokenMemoryCache is a new implementation of TokenStore.

TokenMemoryCache is designed to store up to 2 tokens per server (this is
configurable) for up to 256 servers (this is also configurable), with a
LRU eviction policy. This is so that it works harmoniously with
rustls::ClientSessionMemoryCache, which by default stores resumption
state for up to 256 servers with a LRU eviction policy.

Although this does create a complex intrusive structure, this commit
contains unit tests which assert that it maintains not only the same
external behavior but also isomorphic internal structure to an oracle
implementation based on stdlib types, under a large number of
deterministically randomized operations. This should give us a high
degree of assurance against implementation errors.